feat: Add cast information to useMovie hook
This commit adds the necessary code to the `useMovie` hook to fetch and store the cast information for a movie. The `cast` state is added to the hook and is populated with the cast members using the `getMovieCastUseCase` function. This addition enhances the functionality of the hook by providing access to the cast information for a movie.